refactor: 🔨 로그인 기능 리팩토링

+++ src/main/java/com/daylily/domain/auth/repository/UserRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
     // email은 비어있을 수 있으므로, 유니크 값인 githubId로 조회
-    Optional<User> findByGithubId(Integer githubId);
+    Optional<User> findByGithubId(Long githubId);
 }
